Drop the raw URL path for S3 signing request
See the comment within; the S3 signature needs to use an unescaped
URI path, but in the event where the inbound request already had
an escaped path, the proxied request's signing operates against the
raw form versus the unescaped form.  This forces use of the unescaped
form (again, for the "s3" service only).