Nasty proxy client s3 PUT hacks
This gives us the ability to control header stripping on a per-request
basis using request context stuff (which is probably silly in this case,
but works in the existing code organization without too much trouble).

And it has a PUT to s3 forcibly strip any `Expect: 100-continue` header,
since the S3 API barfs on that.

Furthermore, it forces the outbound request to have a fixed
`Content-Length` header, which means we sidestep net/http.Request's
default of adding a `Transfer-Encoding: chunked` header when the length
is unknown.

This hack is hideous and is for experimentation purpses only.