Get the everything-but-s3 route working
Reworks the signing handlers to use a common base handler since they're
identical save for the actual signing call itself.

Adjusts the main handler routes for the standard signing path to work
properly.  Informally tested with a few misc calls against SQS.

Tweaks the escaping of paths to preserve the leading "/", which is
what AWS signing appears to expect.  This isn't a concern for the path
segment escaper used by S3 since it preserves all "/" instances.