Proper streaming of upstream response
Rearrange the DoProxyRequestHandler to improve some matters:
* Add more informational logging for clarity about request lifecycle
* Immediate response to the originating request with the response status and
  headers
* Stream the response body to the originating request
* Improves throughput, reduces memory footprint, improves the connection
  management (timeouts, monitoring) on the originating client end.

Also tried switching S3 to regular sign (from presign) since we're
loading the request body into memory anyway.  This is almost certainly a
mistake, because...

TODO:
We really should use presigning for S3 PUTs, and avoid the SHA256
calculation entirely, which will have similar operational improvements
for clients to what we get with our response handling (stream the
originating request body to the proxy request body, so the upload
progress perceived by the originating client is an accurate reflection
of the request lifecycle; as it stands, the client sees a very fast
upload and a very slow response, which can lead to timeout issues for
large objects).