Handling PUT body streaming for S3
Similar to the earlier commit that introduced proper response streaming
to the originating client; in the S3 case specifically, it's important
that the PUT request body stream to the canonical endpoint when
possible, so as to avoid originating client socket read timeouts.

This reverts S3 to using presigning, and adds some handlers and
supporting type for allowing the S3 PUT-with-explicit-content-length
case to stream.

Note further that this adds special handling of the empty body case, to
avoid net/http.Request's automatic addition of the `Transfer-Encoding`
header.