Improve S3 header handling
The original proxy implementation signed the proxied request without the
complete set of headers from the original request.  This poses a problem
for S3 because the headers can be fairly rich (for setting various
metadata fields on an S3 object, for instance); the proxy request will
fail if the signature doesn't include all the headers (e.g. the
`X-Amz-Meta-Md5-Content` header).

Changes the handler order so headers are restored for S3 requests prior
to presigning.