New S3-specific path for expect/empty PUTs
Ensures that for an S3 PUT request with an `Expect: 100-continue` header
and an explicit `Content-Length: 0`, we still send a `100-continue`
status line (even though this seems silly); this is consistent with the
upstream S3 behavior and therefore something some clients depend upon.

The routing.go file is absolutely out of control at this point, but I'm
still not gonna clean it up for the time being.