Support explicit non-canonical response headers
The net/http.Response struct gives us headers in the canonicalized form,
which means whatever non-canonical form headers came as from AWS is
lost.  Some clients are sensitive to this in certain cases (like
expecting "ETag" and not "Etag").

Provides a LegacyHeaderMap type to allow different routes to enforce
specific non-canonical forms, and uses it for the S3 cases to enforce
the aforementioned "ETag".